+++ rushub/src/Protocol.h
@@ -52,7 +52,7 @@ public:
 
 	virtual const char * getSeparator() const = 0; ///< protocol separator
 	virtual size_t getSeparatorLen() const = 0; ///< protocol separator length
-	virtual unsigned long getMaxCommandLength() const = 0; ///< protocol max command length
+	virtual unsigned int getMaxCommandLength() const = 0; ///< protocol max command length
 
 	virtual int doCommand(Parser *, Conn *) = 0; ///< doCommand
 	virtual Parser * createParser() = 0; ///< createParser
