+++ rushub/src/Server.h
@@ -71,7 +71,7 @@ public:
 	int mStepDelay; ///< Step delay (for testing)
 	unsigned int mTimerServPeriod; ///< Serv period (msec)
 	unsigned int mTimerConnPeriod; ///< Conn period (msec)
-	unsigned long mMaxSendSize;
+	unsigned int mMaxSendSize;
 	bool mMac; ///< allow to define MAC address
 
 public:
